requirements.txt
PyHamcrest==1.9.0
Django==2.1.5
psycopg2
gunicorn
djangorestframework
reportlab
requests
jsonfield
whitenoise
dj-database-url
scrapy
scrapyd
python-scrapyd-api
scrapy-djangoitem
dill
